+++ lib/cellular_automata.dart
@@ -2,6 +2,7 @@
 library cellular_automata;
 
 export 'package:cellular_automata/src/simulator.dart';
+export 'package:cellular_automata/src/generation.dart';
 export 'package:cellular_automata/src/cell_world.dart';
 export 'package:cellular_automata/src/rules/_ca_rules.dart';
 export 'package:cellular_automata/src/util/array_2d.dart';